requirements.txt
streamlit
openai
python-dotenv
pydantic
jsonschema
cairosvg
httpx
requests
weasyprint
pytest